fix: resolve react hook conflict and type import
