+++ Playproof/src/features/auth/gameInfoPage/components/index.ts
@@ -1,0 +1,1 @@
+export * from './SignupGameInfoForm';
