+++ Playproof/src/features/auth/gameSelectPage/components/SignupGameSelectForm.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/components/ui/Button";
-import { GameSelectGrid } from "@/features/auth/gameSelectPage/components/GameSelectGrid";
+import { GameSelectGrid } from "@/features/auth/gameSelectPage/components";
 import { useSignupGameSelect } from "@/features/auth/gameSelectPage/hooks/useSignupGameSelect";
 
 export function SignupGameSelectForm() {
