+++ Playproof/src/features/auth/gameSelectPage/components/index.ts
@@ -1,0 +1,2 @@
+export * from './GameSelectGrid';
+export * from './SignupGameSelectForm';
