+++ Playproof/src/features/auth/signup/components/index.ts
@@ -1,0 +1,6 @@
+export * from './PasswordSection';
+export * from './PhoneVerificationSection';
+export * from './ProfileSection';
+export * from './SignupCTA';
+export * from './SignupForm';
+export * from './TermsSection';
