+++ Playproof/src/features/matching/components/home/FilteredMatchList.tsx
@@ -1,7 +1,7 @@
 // src/features/matching/components/home/FilteredMatchList.tsx
 import React from 'react';
 import { RefreshCw } from 'lucide-react';
-import { MatchingCard } from '@/features/matching/components/MatchingCard';
+import { MatchingCard } from '@/features/matching/components';
 import type { MatchingData } from '@/features/matching/types';
 
 interface FilteredMatchListProps {
