+++ Playproof/src/features/matching/components/home/PopularMatchList.tsx
@@ -1,7 +1,7 @@
 // src/features/matching/components/home/PopularMatchList.tsx
 import React from 'react';
 import { RefreshCw } from 'lucide-react';
-import { MatchingCard } from '@/features/matching/components/MatchingCard';
+import { MatchingCard } from '@/features/matching/components';
 import type { MatchingData } from '@/features/matching/types';
 
 interface PopularMatchListProps {
