+++ Playproof/src/features/mypage/components/friends/FriendsList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Search } from 'lucide-react';
-import { getFriends, getBlockedUsers, removeFriend, unblockUser } from '../../api/mypageApi';
+import { getFriends, getBlockedUsers, removeFriend, unblockUser } from '@/features/mypage/api/mypageApi';
 import type { FriendData, BlockedUserData } from '@/features/mypage/types';
 import { FriendCard } from './FriendCard';
 import { BlockedUserCard } from './BlockedUserCard';
