+++ Playproof/src/features/profile/components/index.ts
@@ -1,0 +1,5 @@
+export * from './EditProfileForm';
+export * from './GameTagList';
+export * from './ProfileHeader';
+export * from './StatCard';
+export * from './UserProfileModal';
