+++ Playproof/src/features/team/components/azit/RightPanel.tsx
@@ -3,8 +3,7 @@ import React from 'react';
 import type { User, Clip } from '@/types';
 
 // Team 기능 
-import { TeamMemberList } from '@/features/team/components/TeamMemberList';
-import { ClipList } from '@/features/team/components/ClipList';
+import { TeamMemberList, ClipList } from '@/features/team/components';
 
 interface RightPanelProps {
   members: User[];
