+++ Playproof/src/main.tsx
@@ -1,5 +1,4 @@
 //src/main.tsx
-import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
@@ -21,10 +20,7 @@ const queryClient = new QueryClient({
 });
 
 createRoot(document.getElementById("root")!).render(
-	<StrictMode>
-		{/* App을 QueryClientProvider로 감싸주어야 합니다 */}
-		<QueryClientProvider client={queryClient}>
-			<App />
-		</QueryClientProvider>
-	</StrictMode>
+	<QueryClientProvider client={queryClient}>
+		<App />
+	</QueryClientProvider>
 );
